Fix collection loops running indefinitely after reconfig
Previously, collection and log loops would be spawned into the event
loop without a way to cancel them.  If the source was to receive a new
configuration (via the `config` RPC), the old loops would continue
running and report values for an out-of-date config.

Now, we track collection and log loops and cancel them once a new
configuration arrives.

Additionally, long running tasks are decorated with @long_running_task,
which will assign a unique ID to each task and log when they were
started/cancelled.  This should make it easier to see which collection
loops are currently running and help spot any rouge background loops.