Support hosts with mixed record_ids (#26)
This PR adds a workaround for the issue that different hosts use different record_ids for the same entities. In particular, it adds a new config boolean option disable_record_ids. This settings allows to disable the usage of record_ids, allowing the mixed setups with a trade-off to performance.